# backend/internal/apidocs/openapi.yaml
openapi: 3.1.0
info:
  title: Touchline API
  version: 0.1.0
  description: |
    Touchline is a text-first football management game. The backend is a single
    binary (`cmd/touchline serve` runs the API, the scheduler world clock, and
    the event worker in one process; the api/scheduler/worker subcommands split
    them for prod isolation).

    Authentication is cookie-based: login exchanges credentials for a rotating
    access/refresh cookie pair, and every game endpoint below a security scheme
    marker requires a valid `access_token` cookie. The whole world is scoped to
    the session, so "my world" reads are always the caller's own world.

    Launch model: worlds are created as metadata (name), then an admin declares
    countries and leagues, and finally seeds the whole world with clubs,
    players, and memberships. Seasons/league play follow the fixture system
    driven by the scheduler.

    The interactive reference is this page; the raw spec lives at
    `/openapi.yaml`. Live game state (world ticks, match pace 15-min ticks)
    streams over `/ws` (see the WebSocket section).
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development (cmd/touchline serve)
...
paths:
...
  /ws:
    get:
      tags: [realtime]
      summary: Authenticated WebSocket event stream
      operationId: ws
      description: |
        Upgrades to a WebSocket authenticated by the `access_token` cookie on the
        handshake. The server pushes world-scoped events to every socket of that
        world:

        * `world_tick` — `{ event_id, granularity: "daily"|"weekly"|"monthly", tick }`
          broadcast when the scheduler advances the caller's world.
        * `match_update` — live match state updates (`minute`, running score) as the
          worker simulates in-progress matches.

        A HTTP-only cookie is not accessible to the browser WebSocket client
        constructor; Nuxt forwards it automatically because the socket is opened
        same-site. Failing auth closes the handshake with 401.
      security:
        - sessionCookie: []
      responses:
        "101":
          description: WebSocket upgrade accepted; stream begins.
        "401":
          description: Missing/invalid access_token cookie.
  /api/auth/login:
    post:
      tags: [auth]
      summary: Establish a session
      operationId: login
      security: []
      description: |
        Validates credentials and opens a session. An administrator always gets a
        world-less console session. Any other account is resolved per OPD-15(4):

        - the account's active job world wins when it holds one;
        - a jobless account may pass `world_id` to pick a world explicitly (403 if
          it is not a member there);
        - with exactly one joined world the session is minted for that world;
        - with several joined worlds the response is a world PICKER
          `{"status":"worlds","worlds":[...]}` — no cookies are set and the client
          re-posts with the chosen `world_id`;
        - an account with no joined world gets 403 (ErrNoManager).

        The world resolution operates on non-archived memberships.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, example: admin@touchline.local }
                password: { type: string, format: password, example: changeme }
                world_id:
                  type: string
                  format: uuid
                  description: Explicit world pick for a jobless multi-world account (OPD-15(4)(b)).
                  example: 00000000-0000-4000-8000-000000000001
      responses:
        "200":
          description: |
            Either a session (cookies set, `display_name` body) or the world
            picker (`{"status":"worlds","worlds":[...]}`, no cookies).
          headers:
            Set-Cookie:
              description: access_token + refresh_token httpOnly cookies (session outcome only).
              schema: { type: string, example: "access_token=eyJ...; HttpOnly; Path=/; SameSite=Lax; Max-Age=900" }
          content:
            application/json:
              schema:
                oneOf:
                  - title: session
                    type: object
                    properties:
                      display_name: { type: string, example: "Alan Coach" }
                      created_at: { type: string, format: date-time }
                      is_admin: { type: boolean, example: false }
                      id: { type: string, format: uuid }
                  - title: worldPicker
                    type: object
                    required: [status, worlds]
                    properties:
                      status: { type: string, example: worlds }
                      worlds:
                        type: array
                        items: { $ref: "#/components/schemas/WorldChoice" }
        "400":
          description: Malformed body (missing email/password).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "email and password are required" }
        "401":
          description: Invalid credentials.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "invalid email or password" }
        "403":
          description: Account is not loggable into any world (ErrNoManager / ErrNotMember).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                noWorld:
                  value: { error: "no world joined — an administrator must give this account a world before it can log in" }
                notMember:
                  value: { error: "this account is not a member of the requested world" }
  /api/auth/register:
    post:
      tags: [auth]
      summary: Create a manager account (signup)
      operationId: register
      security: []
      description: |
        Product signup flow (A13). Creates a plain (non-admin) account and — when
        there is EXACTLY one playable world (`active`/`open_beta`) — auto-joins it
        as an unemployed manager and auto-issues a first job offer from the first
        available AI club (deterministic by club id). With zero or several
        playable worlds the account is created world-less (`world: null`) and an
        administrator must join it later. Auto-offer is best-effort: when no AI
        club is available the account is created and `offer` is null. No session
        is minted by registration; the new manager logs in via /api/auth/login.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, example: player@touchline.local }
                password: { type: string, format: password, example: changeme }
                display_name:
                  type: string
                  description: Defaults to the email local part.
                  example: Alan Coach
      responses:
        "201":
          description: Account created (and joined + offered when possible).
          content:
            application/json:
              schema:
                type: object
                required: [id, email, display_name, is_admin, world, offer]
                properties:
                  id: { type: string, format: uuid }
                  email: { type: string, format: email }
                  display_name: { type: string }
                  is_admin: { type: boolean, enum: [false] }
                  world:
                    description: The joined world, or null when no auto-join happened.
                    allOf:
                      - $ref: "#/components/schemas/WorldChoice"
                      - nullable: true
                  offer:
                    description: The auto-issued first job offer, or null.
                    $ref: "#/components/schemas/JobOffer"
              examples:
                joined:
                  value:
                    id: 00000000-0000-4000-8000-000000000100
                    email: player@touchline.local
                    display_name: Alan Coach
                    is_admin: false
                    world:
                      world_id: 00000000-0000-4000-8000-000000000001
                      name: Westland
                      status: active
                    offer:
                      id: 00000000-0000-4000-8000-000000000009
                      world_id: 00000000-0000-4000-8000-000000000001
                      club_id: 00000000-0000-4000-8000-000000000005
                      club_name: Northgate Rovers
                      manager_id: 00000000-0000-4000-8000-000000000101
                      status: proposed
                      created_at: "2026-01-01T00:00:00Z"
                worldless:
                  value:
                    id: 00000000-0000-4000-8000-000000000100
                    email: player@touchline.local
                    display_name: Alan Coach
                    is_admin: false
                    world: null
                    offer: null
        "400":
          description: Malformed body (missing email/password).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "email and password are required" }
        "409":
          description: An account with this email already exists (ErrEmailTaken).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "an account with this email already exists" }
  /api/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate the session
      operationId: refresh
      security: []
      description: |
        Exchanges the `refresh_token` cookie for a fresh cookie pair for the same
        manager. Used when the access token expires.
      responses:
        "200":
          description: Session rotated; new cookie pair set.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
              example: { status: ok }
        "401":
          description: Missing/invalid refresh token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "missing refresh token" }
  /api/dashboard:
    get:
      tags: [dashboard]
      summary: Home dashboard aggregate (S07-01)
      operationId: dashboard
      description: |
        The manager's home feed: urgent, important and interesting items
        aggregated from the world's already-persisted state (open bid threads,
        expiring contracts, imminent fixtures, board confidence, finances,
        player morale, table position, rival results, market news). Postgres is
        authoritative; the same items are pushed incrementally over the socket
        as `dashboard_update` events. A manager with no active club gets empty
        sections rather than an error.
      responses:
        "200":
          description: Aggregated dashboard feed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DashboardSnapshot" }
        "401":
          description: No valid access_token cookie.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "unauthorized" }
        "403":
          description: The caller has no world context.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "no world context" }
  /api/news:
    get:
      tags: [dashboard]
      summary: Manager news feed
      operationId: newsFeed
      description: |
        The world's published news stories, newest first, scoped to the
        calling manager's world (resolved from the session identity).
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum items returned, capped at 100.
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: The manager's world news feed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  world_id: { type: string, format: uuid }
                  stories:
                    type: array
                    items: { $ref: "#/components/schemas/NewsStory" }
        "401":
          description: No valid access_token cookie.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "unauthorized" }
        "403":
          description: The caller has no world context.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "no world context" }
  /api/clubs:
    get:
      tags: [clubs]
      summary: List clubs in my world
      operationId: listClubs
      description: Clubs of the caller's own world (AI clubs included).
      responses:
        "200":
          description: Club list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  clubs:
                    type: array
                    items: { $ref: "#/components/schemas/Club" }
              examples:
                sample:
                  value:
                    clubs:
                      - id: 00000000-0000-4000-8000-000000000005
                        world_id: 00000000-0000-4000-8000-000000000001
                        name: Northgate Rovers
                        short_name: NGR
        "403":
          description: Session has no world context (admin console session).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "no world context" }
  /api/clubs/{id}:
    get:
      tags: [clubs]
      summary: Get a club in my world
      operationId: getClub
      parameters:
        - $ref: "#/components/parameters/ClubID"
      responses:
        "200":
          description: Club detail with full squad.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClubDetail" }
              examples:
                sample:
                  value:
                    id: 00000000-0000-4000-8000-000000000005
                    world_id: 00000000-0000-4000-8000-000000000001
                    name: Northgate Rovers
                    short_name: NGR
                    country: England
                    is_ai_controlled: false
                    manager:
                      id: 00000000-0000-4000-8000-000000000006
                      display_name: Test Manager
                    squad:
                      - id: 00000000-0000-4000-8000-00000000000b
                        person_id: 00000000-0000-4000-8000-00000000000c
                        first_name: Leo
                        last_name: Marsh
                        display_name: Leo Marsh
                        nationality_code: ENG
                        nationality_name: England
                        date_of_birth: "1998-04-12T00:00:00Z"
                        age: 27
                        primary_position: ST
                        squad_number: 9
        "400":
          description: Invalid club id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "invalid club id" }
        "403":
          description: No world context on the session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Club not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "club not found" }
  /api/clubs/{id}/lineup:
    get:
      tags: [clubs]
      summary: Get the club's lineup
      operationId: getLineup
      description: Slot -> player mapping (16 slots; empty slots null).
      parameters:
        - $ref: "#/components/parameters/ClubID"
      responses:
        "200":
          description: The current lineup.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lineup" }
              examples:
                sample:
                  value:
                    slots:
                      - slot: 0
                        player_id: 00000000-0000-4000-8000-00000000000b
    put:
      tags: [clubs]
      summary: Set the club's lineup
      operationId: setLineup
      description: Submit a full 16-slot lineup. Requires the club's manager identity (or an AI policy bot in scheduled mode).
      parameters:
        - $ref: "#/components/parameters/ClubID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slots]
              properties:
                slots:
                  type: array
                  minItems: 16
                  maxItems: 16
                  items:
                    type: object
                    required: [slot]
                    properties:
                      slot: { type: integer, example: 0 }
                      player_id: { type: string, format: uuid, nullable: true, example: "00000000-0000-4000-8000-00000000000b" }
      responses:
        "204":
          description: Lineup saved.
        "400":
          description: Invalid lineup (players not in squad, duplicated, wrong count).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthenticated.
        "403":
          description: Not the owning manager of the club.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Club not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/clubs/{id}/tactics:
    get:
      tags: [clubs]
      summary: Get the club's tactics
      operationId: getTactics
      parameters:
        - $ref: "#/components/parameters/ClubID"
      responses:
        "200":
          description: Style, formation, allowed formations, and current slot assignments.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tactics" }
              examples:
                sample:
                  value:
                    club_id: 00000000-0000-4000-8000-000000000005
                    style: balanced
                    formation: 4-3-3
                    allowed_formations: ["4-3-3", "4-4-2", "3-5-2"]
                    slots:
                      - slot: 0
                        position: GK
                        player_id: 00000000-0000-4000-8000-00000000000b
    post:
      tags: [clubs]
      summary: Set the club's tactics
      operationId: setTactics
      description: Change playing style and/or formation (effect on the existing slot mapping is a bulk re-check by the service).
      parameters:
        - $ref: "#/components/parameters/ClubID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [style, formation]
              properties:
                style: { type: string, enum: [balanced, counter, attacking, defensive, press, possession], example: balanced }
                formation: { type: string, example: "4-3-3" }
      responses:
        "204":
          description: Tactics saved.
        "400":
          description: Invalid formation/style for the squad.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Not the owning manager.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Club not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/clubs/{id}/training-plan:
    get:
      tags: [clubs]
      summary: Get the club's training plan
      operationId: getTrainingPlan
      parameters:
        - $ref: "#/components/parameters/ClubID"
      responses:
        "200":
          description: The club's weekly training archetype.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrainingPlan" }
              examples:
                sample:
                  value:
                    club_id: 00000000-0000-4000-8000-000000000005
                    archetype: balanced
                    effective_from_tick: 62
                    last_applied_week: 7
    post:
      tags: [clubs]
      summary: Set the club's training plan
      operationId: setTrainingPlan
      parameters:
        - $ref: "#/components/parameters/ClubID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [archetype]
              properties:
                archetype:
                  type: string
                  enum: [balanced, attacking, defending, fitness, technical]
                  example: balanced
      responses:
        "204":
          description: Training plan saved.
        "400":
          description: Unknown archetype.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Not the owning manager.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/clubs/{id}/finances:
    get:
      tags: [finance]
      summary: Club financial summary
      operationId: getFinances
      description: Owning manager only (the club resolves the owner's world).
      parameters:
        - $ref: "#/components/parameters/ClubID"
      responses:
        "200":
          description: Cash, budgets, wages, and projections.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FinancialSummary" }
              examples:
                sample:
                  value:
                    currency: GBP
                    cash: 48250000
                    operating_profit: 740000
                    transfer_budget: { season: 1, allocated: 30000000, committed: 0, available: 30000000 }
                    wage_budget: { season: 1, allocated: 18000000, committed: 16400000, available: 1600000 }
                    wage_commitments: { count: 24, weekly_wage: 316000, annual_wage: 16400000 }
                    committed_spending: 0
                    projected_revenue: 125000000
                    projected_year_end_balance: 30400000
                    future_installments: 0
                    debt: 0
                    factors:
                      - { label: League position, amount: 18000000 }
        "403":
          description: Not the owning manager.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Club not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/transfers/listings/{id}:
    get:
      tags: [transfer]
      summary: Listing detail
      operationId: getListing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid, example: "00000000-0000-4000-8000-000000000020" }
      responses:
        "200":
          description: The listing with its open bids.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Listing" }
        "404":
          description: Listing not found (or in another world).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: httpOnly cookie set by POST /api/auth/login and rotated by /api/auth/refresh.
...
  schemas:
...
    WorldChoice:
      type: object
      description: One joined world offered by the login world picker / registration join.
      required: [world_id, name, status]
      properties:
        world_id: { type: string, format: uuid }
        name: { type: string }
        status: { type: string, enum: [provisioning, seeding, active, open_beta, paused, archived] }
...
security:
  - sessionCookie: []